Add scripts to rule inputs explicitely
This whole thing of rerunning rules automatically when scripts have changed doesnt work well for Snakemake. Instead, I am doing it manually and explicitely here by adding the scripts to the rule's inputs explicitely. Of course this won't work for rules without scripts but with shell inline scripts, but it's the best solution I see at the moment.