Fix dont build docs through Snakemake
Now that we need eurocalliopelib for building the docs, the env does not work for both: Snakemake execution and normal execution. That is because paths in Snakemake execution get resolved differently.

We could have two different envs for the two purposes. But in fact, it's not necessary to build the docs through Snakemake. It's a simple one-liner, and it's well known to everyone knowing mkdocs (and we will document it).