Fix Linux compatibility with netcdf
On Linux, there seems to be a problem with compatibilities of netcdf libraries that is not considered by the conda packages. The libraries seem to be libnetcdf and hdf5. Here, I am pinning these libraries to a combination that works with Linux, ARM Mac, and all other dependencies we have to introduce the smallest change possible.